fix: pin scientific-notation guard, correct NumberLiteral's type, widen dict_in_dict range
Three review findings on the decimal-render commit:

- _number's switch to format(value, "f") was unpinned -- str(value)
  passed the whole suite. Added a direct test for
  Decimal("0.0000001"), whose str() is "1E-7" (does not re-lex) but
  whose format(_, "f") is "0.0000001". Verified it discriminates:
  reverting to str(value) fails exactly this one test in 2037.

- nodes.py's NumberLiteral.value was still annotated int, the
  opposite of the real invariant -- a bare int renders as
  "5.000000" with no error anywhere. Corrected to Decimal with a
  comment explaining why. That exposed test_treeview.py's generic
  _minimal() helper, which pattern-matches field types and had no
  Decimal branch; taught it one, per its own assertion's
  instruction, and updated its two NumberLiteral(1) call sites to
  match.

- test_the_generator_produces_the_dictionary_shapes_too's
  dict_in_dict shape fell 7 -> 4 in the same _NUMBERS reshuffle that
  cost unplug_under_eq its only hit, with its last surviving seed
  landing on the final seed of the 300-seed range. Widened to 600
  (9 hits, last one 17 seeds shy of the new boundary) with a comment
  recording both seed sets honestly, not a change-probability claim.

+++ src/matrixlang/nodes.py
@@ -10,6 +10,7 @@ different line/column numbers, so positions carry compare=False.
 """
 
 from dataclasses import dataclass, field
+from decimal import Decimal
 
 from matrixlang.tokens import TokenType
 
@@ -30,7 +31,13 @@ class Expr(Node):
 
 @dataclass
 class NumberLiteral(Expr):
-    value: int
+    # Must be a Decimal, not a bare int -- this is load-bearing, not a
+    # hint. render.py's _number walks format(value, "f") character by
+    # character; handed a plain int, format(5, "f") silently succeeds as
+    # "5.000000" instead of raising, so a stray int here renders wrong
+    # (`trace 5.000000`) with no error anywhere, and the interpreter
+    # only catches the mistake later, at arithmetic.
+    value: Decimal
 
 
 @dataclass
